refactor: useSidebar hook 분리
